refactor : 알림 기능에 옵저버 디자인 패턴 적용
