--- lime-api/src/main/java/com/programmers/lime/domains/sse/IAlarmSubject.java
@@ -1,12 +1,0 @@
-package com.programmers.lime.domains.sse;
-
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-
-public interface IAlarmSubject {
-
-	Map<Long, IAlarmObserver> observers = new ConcurrentHashMap<>();
-
-	void registerObserver(Long receiverId,IAlarmObserver o);
-	void notifyObserver(SsePayload ssePayload);
-}
